make 'docker compose restart api' work.

+++ container_init/api/api-entrypoint.bash
@@ -1,2 +1,5 @@
-ln -sF /proc/1/fd/1 /var/log/phaidra/api.log
+if [[ ! -f /var/log/phaidra/api.log ]]
+then
+    ln -sF /proc/1/fd/1 /var/log/phaidra/api.log
+fi
 hypnotoad -f phaidra-api.cgi
